Tests/LibWeb/Text/expected/form-formEnctype-attribute.txt

button: unset
button.getAttribute('formEnctype') == 'null'
button.formEnctype == ''

button.setAttribute('formEnctype', '')
button.getAttribute('formEnctype') == ''
button.formEnctype == 'application/x-www-form-urlencoded'

button.setAttribute('formEnctype', 'undefined')
button.getAttribute('formEnctype') == 'undefined'
button.formEnctype == 'application/x-www-form-urlencoded'

button.setAttribute('formEnctype', 'null')
button.getAttribute('formEnctype') == 'null'
button.formEnctype == 'application/x-www-form-urlencoded'

button.setAttribute('formEnctype', 'application/x-www-form-urlencoded')
button.getAttribute('formEnctype') == 'application/x-www-form-urlencoded'
button.formEnctype == 'application/x-www-form-urlencoded'

button.setAttribute('formEnctype', 'multipart/form-data')
button.getAttribute('formEnctype') == 'multipart/form-data'
button.formEnctype == 'multipart/form-data'

button.setAttribute('formEnctype', 'text/plain')
button.getAttribute('formEnctype') == 'text/plain'
button.formEnctype == 'text/plain'

button.setAttribute('formEnctype', 'APPLICATION/X-WWW-FORM-URLENCODED')
button.getAttribute('formEnctype') == 'APPLICATION/X-WWW-FORM-URLENCODED'
button.formEnctype == 'application/x-www-form-urlencoded'

button.setAttribute('formEnctype', 'MULTIPART/FORM-DATA')
button.getAttribute('formEnctype') == 'MULTIPART/FORM-DATA'
button.formEnctype == 'multipart/form-data'

button.setAttribute('formEnctype', 'tEXt/PlAIn')
button.getAttribute('formEnctype') == 'tEXt/PlAIn'
button.formEnctype == 'text/plain'

button.setAttribute('formEnctype', 'text/plain ')
button.getAttribute('formEnctype') == 'text/plain '
button.formEnctype == 'application/x-www-form-urlencoded'

button.setAttribute('formEnctype', '7')
button.getAttribute('formEnctype') == '7'
button.formEnctype == 'application/x-www-form-urlencoded'

button.setAttribute('formEnctype', '5%')
button.getAttribute('formEnctype') == '5%'
button.formEnctype == 'application/x-www-form-urlencoded'
